Move scheduling constants and task model
